Bug Fix - Source Handler
